Update metrics page to show chart if no data is available
